pyproject.toml
[project]
name = "ia-para-desarrolladores"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.9",
    "bs4>=0.0.2",
    "chromadb>=1.0.13",
    "crewai>=0.134.0",
    "deprecated>=1.2.18",
    "fastapi>=0.115.13",
    "fastapi-mcp>=0.3.4",
    "fastmcp>=2.9.0",
    "google-adk>=1.1.1",
    "google-genai==1.16.1",
    "google-generativeai==0.8.5",
    "gradio>=5.34.2",
    "ipykernel==6.29.5",
    "kaggle>=1.7.4.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-experimental>=0.3.4",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai==2.0.10",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "markdown2>=2.5.3",
    "mcp[cli]>=1.9.4",
    "numexpr>=2.11.0",
    "openai>=1.91.0",
    "openai-agents>=0.0.19",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20250506",
    "pi-heif>=0.22.0",
    "pillow-heif>=0.22.0",
    "playwright>=1.52.0",
    "plotnine==0.14.5",
    "pymupdf>=1.26.1",
    "pypandoc>=1.15",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "scipy==1.15.3",
    "statsmodels==0.14.4",
    "tabulate>=0.9.0",
    "torch>=2.7.1",
    "unstructured>=0.17.2",
    "unstructured-inference>=1.0.5",
    "unstructured-pytesseract>=0.3.15",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
    "yfinance>=0.2.64",
]
